+++ changelog.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 - Require NodeJS 18
 
+### Changed
+- Use native `fetch` instead of superagent library
+
 ## [6.0.0] – 2023-04-08
 
 ### Breaking